+++ internal/redfish/client_test.go
@@ -3,7 +3,7 @@ package redfish
 import (
 	"testing"
 
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
+	clusterv1 "sigs.k8s.io/cluster-api/api/core/v1beta2"
 )
 
 func TestConvertToMachineAddresses(t *testing.T) {